+++ src/main/java/frc/robot/commands/subsystems/arm/SetArmCommand.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems.arm.commands;
+package frc.robot.commands.subsystems.arm;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Presets.Preset;
